backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
regex==2023.10.3
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
python-dotenv==1.0.0